feat: construct kernel version of Pratt parser from scratch
This uses Eli's code as a model, but is meant to be worked from the
ground up, so that I can understand it.

So far, this only evaluates expressions consisting of a single
literal. Also, the input is assumed to already come in tokenized form,
an almost universal assumption in examples on the Web.