feat: enable names as local variables
These take the place of registers: when a new name is declared,
another register space is added.

Technically, the old system of only using numerical values for address
will still work; however, this is a meaningful sugar on top of
that. Hopefully, it would be possible to eliminate the possiblity of
using "raw" addresses.

There are two issues though:

1. Local variables can be redeclared. In such a case, the old
declaration is the one that stays in use.

2. There's no check against using variables that haven't been declared
yet.