feat: define ';' nud behavior as a no-op
As a nud, i.e. prefix operator ';' is a no-op. This lets users input
empty "statements" like ';;', because all the empty semicolons that
occur before the last one are merely discarded by the evaluator.

This also lets a future preprocessing step inject semicolons where
needed; for example, we may want to replace newlines with semicolons
inside source files. However, we want to do this as flatly and easily
as possible, and so for example inserting a semicolon after a comment
should cause no harm.